Listado de usuarios de ejemplo
- Apretar el botón de solicitud de amistad (por el momento y para testear)
muestra una lista de todos los usuarios que se encuentren en el sistema con
ids entre 1 y 20.
- UserList funcionando, y tocar un resultado abre el usuario respectivo.
Falta corregir un par de cosas como un iconito de like para las recomendacio-
nes y el ícono de arriba a la izquierda que por alguna razón es una flecha,
como también el título en la barra.
- NavDrawer más generalizado. NavDrawer en UserListActivity.

Closes #19

+++ app/Jobify/app/src/main/java/ar/fiuba/jobify/app_server_api/User.java
@@ -91,7 +91,7 @@ public class User {
         return workHistory;
     }
 
-    public int getCantRecomendaciones() {   // TODO
+    public long getCantRecomendaciones() {   // TODO
         return 123;
         //return cantRecomendaciones;
     }
